+++ NEWS.md
@@ -104,7 +104,7 @@
   cfact_hist (for historical counterfactuals), cfact_fore (for counterfactual forecast scenarios), and cfact_girf (for counterfactual
   generalized impulse response functions). See the vignette for details on the implemented methods.
 * MAJOR: Added the new function hist_decomp that allows to compute historical decompositions for TVAR and STVAR models. See the vignette for details.
-* It is not possible in the genetic algorithm to only allow for estimates that allocate the specified amount of observations to each regime
+* It is not possible in the genetic algorithm to only allow for estimates that allocate a specified amount of observations to each regime
   (see ??GAfit and the arguments bound_by_weights and min_obs_coef_ga).
 * Bug fix: There was an issue with the Phase 1 estimation of the three-phase estimation when weight_function = "exogenous" (NLS estimates were not calculated
   correctly). This is now fixed.
@@ -118,3 +118,4 @@
 
 * Fixed the labels for variables in GIRFs etc when they were calculated using a model either not containing any data or containing data without variable names. 
 * Removed the fixed "lwd" setting from the plot method for historical decompositions. The line thickness can now be adjust with the dot parameters. 
+* If a matrix is provided as the argument "init_vals" in GIRF or GFEVD, it is now automatically converted to the appropriate array.